Extended shop to overwrite price by `=` option
The <form> and <option> were extended by data attributes which can be used to parse it by JavaScript (for whatever reason).
Before the option could be prefixed by `-` and `+` to subtract/add a sum from the default product price. With prefix `=` the product price will be overwritten by the selected option.
In addition there is a new template option to hide the price option.